feat(app:asctb-reporter): :sparkles: ported functionality from root component to app component to make drawers appear outside the router
