+++ apps/asctb-reporter/src/app/components/indented-list/indented-list.component.ts
@@ -1,4 +1,5 @@
 import { FlatTreeControl } from '@angular/cdk/tree';
+import { CommonModule } from '@angular/common';
 import {
   AfterViewInit,
   Component,
@@ -10,9 +11,13 @@ import {
   ViewChild,
   inject,
 } from '@angular/core';
-import { MatTree, MatTreeFlatDataSource, MatTreeFlattener } from '@angular/material/tree';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTree, MatTreeFlatDataSource, MatTreeFlattener, MatTreeModule } from '@angular/material/tree';
 import { ILNode } from '../../models/indent.model';
 import { Row, Sheet } from '../../models/sheet.model';
+import { SidenavHeaderModule } from '../sidenav-header/sidenav-header.module';
+import { SidenavModule } from '../sidenav/sidenav.module';
 import { IndentedListService } from './indented-list.service';
 
 interface Node {
@@ -30,9 +35,9 @@ interface FlatNode {
 
 @Component({
   selector: 'app-indent',
+  imports: [CommonModule, SidenavModule, SidenavHeaderModule, MatTreeModule, MatButtonModule, MatIconModule],
   templateUrl: './indented-list.component.html',
   styleUrls: ['./indented-list.component.scss'],
-  standalone: false,
 })
 export class IndentedListComponent implements OnInit, OnDestroy, AfterViewInit {
   readonly indentService = inject(IndentedListService);
