+++ apps/asctb-reporter/src/app/modules/navbar/navbar.component.html
@@ -1,4 +1,13 @@
 <div class="toolbar-nav">
+  <div header>
+    <div class="branding">
+      <hra-icon [svgIcon]="'product:asctb-reporter'"></hra-icon>
+      <div class="app-title">ASCT+B Reporter</div>
+    </div>
+    <button mat-icon-button hraIconButtonSize="large" hraIconButtonVariant="dark" [matMenuTriggerFor]="headerMenu">
+      <mat-icon>more_vert</mat-icon>
+    </button>
+  </div>
   <div class="nav-wrapper">
     <div class="nav-item">
       <app-search [disabled]="(ui$ | async)?.error?.hasError ?? false"></app-search>
@@ -12,19 +21,7 @@
       </mat-form-field>
     </div>
 
-    <!-- <app-nav-item
-      #tooltip="matTooltip"
-      matTooltip="Master Data Tables"
-      matTooltipPosition="below"
-      [label]="'Data Tables'"
-      [matMenuTriggerFor]="nestedMenu.childMenu"
-    >
-      <mat-icon class="toolbar-icons">table_view</mat-icon>
-      <app-table-nested-menu
-        #nestedMenu
-        [sheetDetails]="(configService.allSheetConfigurations$ | async) ?? []"
-      ></app-table-nested-menu>
-    </app-nav-item>
+    <!--
 
     <app-nav-item
       [label]="'Refresh'"
@@ -34,16 +31,6 @@
       matTooltipPosition="below"
     >
       <mat-icon class="toolbar-icons">refresh</mat-icon>
-    </app-nav-item>
-
-    <app-nav-item
-      [label]="'Debug Log'"
-      (click)="toggleDebugLogs()"
-      #tooltip="matTooltip"
-      matTooltip="Debug Log"
-      matTooltipPosition="below"
-    >
-      <mat-icon class="toolbar-icons">bug_report</mat-icon>
     </app-nav-item>
 
     <app-nav-item
@@ -81,3 +68,14 @@
     </mat-menu> -->
   </div>
 </div>
+
+<mat-menu #headerMenu="matMenu" class="header-menu">
+  <button mat-menu-item (click)="togglePane()">
+    <mat-icon>visibility_{{ controlPaneOpen() ? 'off' : 'on' }}</mat-icon>
+    {{ controlPaneOpen() ? 'Hide' : 'Show' }} legend & visualization controls
+  </button>
+  <button mat-menu-item (click)="toggleDebugLogs()">
+    <mat-icon>bug_report</mat-icon>
+    Debug log
+  </button>
+</mat-menu>
