+++ apps/asctb-reporter/src/app/modules/navbar/navbar.component.scss
@@ -1,7 +1,28 @@
 @use 'vars';
+@use 'utils';
 
 :host {
   display: block;
+
+  div[header] {
+    border-right: 1px solid vars.$outline;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0.5rem 1rem;
+    height: 5rem;
+    width: 20.5rem;
+
+    .branding {
+      display: flex;
+      align-items: center;
+      gap: 1rem;
+
+      .app-title {
+        @include utils.use-font(title, medium);
+      }
+    }
+  }
 
   .nav-item {
     flex: 1;
@@ -18,7 +39,7 @@
     display: flex;
     padding: 1rem;
     gap: 0.625rem;
-    width: 100%;
+    flex: 1;
     align-items: center;
   }
