+++ apps/asctb-reporter/src/app/modules/root/root.component.scss
@@ -1,7 +1,7 @@
 @use 'vars';
 
 .control-pane-container {
-  height: calc(100% - 7.1rem);
+  height: calc(100% - 12.1rem);
   width: 100%;
   position: absolute;
   left: 0;
